feat: refactor pricing section and introduce PricingPlan component for improved structure and clarity
